Removed components
Removed Recipe, Ingredient, and Admin Components. Moved original admin functionality to Recipes (plural) component and will now use that to manage Recipes and Ingredients
