+++ src/app/app.component.html
@@ -11,14 +11,12 @@
               Admin
             </a>
             <ul class="dropdown-menu" aria-labelledby="navbarDropdown">
-              <li><a class="dropdown-item" href="/admin">Recipes and Ingredients</a></li>
+              <li><a class="dropdown-item" href="/recipes">Recipes and Ingredients</a></li>
               <li><a class="dropdown-item" href="/users">Users CMS</a></li>
               <li><a class="dropdown-item" href="/locations">Location Manager</a></li>
             </ul>
           </li>
-          <li class="nav-item">
-            <a class="nav-link active" aria-current="page" href="/admin">Admin</a>
-          </li>
+
           <li class="nav-item">
             <a class="nav-link active" aria-current="page" href="/users">Users</a>
           </li>
