--- src/app/views/recipe/recipe.component.html
@@ -1,21 +1,0 @@
-
-<div class="container">
-    <div class="row">
-        <div class="col-12">
-            <h5>Recipe Name</h5>
-            <input type="text" pInputText [(ngModel)]="recipe.name"/>
-        </div>
-
-        <div class="col-6">
-            <h5>Estimated Time (in min.)</h5>
-            <input type="text" pInputText [(ngModel)]="recipe.minutes"/>        
-        </div>
-
-        <div class="col-6">
-            <h5>Estimated Time (in min.)</h5>
-            <input type="text" pInputText [(ngModel)]="recipe.minutes"/>        
-        </div>
-    </div>
-
-
-</div>
